move check dataframe to auto estimator
