+++ python/orca/src/bigdl/orca/automl/auto_estimator.py
@@ -183,6 +183,7 @@ class AutoEstimator:
                 "This AutoEstimator has already been fitted and cannot fit again.")
 
         metric_mode = AutoEstimator._validate_metric_mode(metric, metric_mode)
+        AutoEstimator._check_spark_dataframe_input(data, validation_data, feature_cols, target_cols)
 
         self.searcher.compile(data=data,
                               model_builder=self.model_builder,
@@ -258,3 +259,29 @@ class AutoEstimator:
         if mode not in ["min", "max"]:
             raise ValueError("`mode` has to be one of ['min', 'max']")
         return mode
+
+    @staticmethod
+    def _check_spark_dataframe_input(data,
+                                     validation_data,
+                                     feature_cols,
+                                     target_cols
+                                     ):
+
+        def check_cols(cols, cols_name):
+            if not cols:
+                raise ValueError(f"You must input valid {cols_name} for Spark DataFrame data input")
+            if isinstance(cols, list):
+                return cols
+            if not isinstance(cols, str):
+                raise ValueError(f"{cols_name} should be a string or a list of strings, "
+                                 f"but got {type(cols)}")
+            return [cols]
+
+        from pyspark.sql import DataFrame
+        if isinstance(data, DataFrame):
+            feature_cols = check_cols(feature_cols, cols_name="feature_cols")
+            target_cols = check_cols(target_cols, cols_name="target_cols")
+            if validation_data:
+                if not isinstance(validation_data, DataFrame):
+                    raise ValueError(f"data and validation_data should be both Spark DataFrame, "
+                                     f"but got validation_data of type {type(data)}")
